fix: sıkıldım artık hadi

+++ database/migrations/2024_05_09_124712_create_inquiries_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('inquiries', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('treatment_id');
+            $table->integer('treatment_id');
             $table->integer('status')->default(\App\Enums\InquiryStatus::WAITING->value); //onaylandı, reddedildi, bekliyor,
             $table->integer('gender')->default(\App\Enums\Gender::None->value);
